Use `jj bookmark move` to move to avoid accidental issues with main | master
If we use `jj bookmark set` it will effectively run a create or update
operation. For this specific example, though, we expect a preexisting
bookmark to exist. Since it's really common for me to have some repos
that use `master` and others that use `main` this distinction is really
important.

tldr; if you use `bookmark set master -r @-` on a repo that is using
`main` as the primary branch you haven't at all done what you wanted
(which is to move up `main` to that change) you've just made a new
bookmark that is untracked and unpushed.